create_jwt

+++ src/main/java/com/example/msl_b/Controller/PostController.java
@@ -1,59 +1,63 @@
 package com.example.msl_b.Controller;
 
-import com.example.msl_b.Domain.Entity.Post;
 import com.example.msl_b.Domain.Entity.User;
 import com.example.msl_b.Domain.Repository.UserRepository;
 import com.example.msl_b.Domain.dto.PostDto;
+import com.example.msl_b.config.auth.PrincipalDetails;
 import com.example.msl_b.exception.Response;
 import com.example.msl_b.serivce.PostService;
-import com.example.msl_b.serivce.UserService;
 import io.swagger.annotations.ApiOperation;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
+
 import java.math.BigInteger;
-import java.util.List;
 
 @RestController
 @AllArgsConstructor
 public class PostController {
 
+
     private final PostService postService;
-    private final UserService userService;
     private final UserRepository userRepository;
 
+
+    // 전체 게시글 조회
     @ApiOperation(value = "전체 게시글 보기", notes = "전체 게시글을 조회한다.")
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/posts")
-    public List search(@RequestBody PostDto postDto) {
-
-        List<PostDto> boardList = postService.getBoardlist();
-        return postService.getBoardlist();
+    @GetMapping("/boards")
+    public Response getBoards() {
+        return new Response("성공", "전체 게시물 리턴", postService.getBoards());
     }
 
+
+    // 개별 게시글 조회
     @ApiOperation(value = "개별 게시글 보기", notes = "개별 게시글 조회한다.")
     @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/post/{id}") // 상세 조회
-    public Post get(@PathVariable("id") BigInteger id) {
-        return postService.findOne(id);
+    @GetMapping("/boards/{id}")
+    public Response getBoard(@PathVariable("id") BigInteger id) {
+        return new Response("성공", "개별 게시물 리턴", postService.getBoard(id));
     }
 
+
+    // 게시글 작성
     @ApiOperation(value = "게시글 작성", notes = "게시글을 작성한다.")
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/boards/write")
-    public Response write(@RequestBody PostDto postDto) {
-        // 원래 로그인을 하면, User 정보는 세션을 통해서 구하고 주면 되지만,
-        // 지금은 핵심 개념을 알기 위해서, JWT 로그인은 생략하고, 임의로 findById 로 유저 정보를 넣어줬습니다.
-
-        User user = userRepository.findById(1).get();
-        return new Response("성공", "글 작성 성공", boardService.write(boardDto, user));
+    public Response write(@RequestBody PostDto boardDto, Authentication authentication) {
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        return new Response("성공", "글 작성 성공", postService.write(boardDto, user));
     }
 
+
+    // 게시글 수정
     @ApiOperation(value = "게시글 수정", notes = "게시글을 수정한다.")
     @ResponseStatus(HttpStatus.OK)
     @PutMapping("/boards/update/{id}")
-    public Response edit(@RequestBody BoardDto boardDto, @PathVariable("id") Integer id) {
+    public Response edit(@RequestBody PostDto boardDto, @PathVariable("id") BigInteger id, Authentication authentication) {
         // 원래 로그인을 하면, User 정보는 세션을 통해서 구하고 주면 되지만,
         // 지금은 핵심 개념을 알기 위해서, JWT 로그인은 생략하고, 임의로 findById 로 유저 정보를 넣어줬습니다.
 
@@ -64,19 +68,34 @@ public class PostController {
         // 맞으면 아래 로직 수행, 틀리면 다른 로직(ResponseFail 등 커스텀으로 만들어서) 수행
         // 이건 if문으로 처리할 수 있습니다. * 이 방법 말고 service 내부에서 확인해도 상관 없음
 
-        User user = userRepository.findById(1).get();
-        return new Response("성공", "글 수정 성공", boardService.update(id, boardDto));
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        if (user.getName().equals(postService.getBoard(id).getWriter())) {
+            // 로그인된 유저의 글이 맞다면
+            return new Response("성공", "글 수정 성공", postService.update(id, boardDto));
+        } else {
+            return new Response("실패", "본인 게시물만 수정할 수 있습니다.", null);
+
+        }
     }
 
 
+    // 게시글 삭제
     @ApiOperation(value = "게시글 삭제", notes = "게시글을 삭제한다.")
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/boards/delete/{id}")
-    public Response delete(@PathVariable("id") BigInteger id) {
+    public Response delete(@PathVariable("id") BigInteger id, Authentication authentication) {
         // 이것도 마찬가지로, JWT(로그인 관련) 공부를 하고나서 현재 이 요청을 보낸 로그인된 유저의 정보가
         // 게시글의 주인인지 확인하고, 맞으면 삭제 수행 후 리턴해주고, 틀리면 에러 리턴해주면 됩니다.
 
-        postService.delete(id);
-        return new Response("성공", "글 삭제 성공", null);
+        PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
+        User user = principalDetails.getUser();
+        if (user.getName().equals(postService.getBoard(id).getWriter())) {
+            // 로그인된 유저가 글 작성자와 같다면
+            postService.delete(id); // 이 메소드는 반환값이 없으므로 따로 삭제 수행해주고, 리턴에는 null을 넣어줌
+            return new Response("성공", "글 삭제 성공", null);
+        } else {
+            return new Response("실패", "본인 게시물만 삭제할 수 있습니다.", null);
+        }
     }
 }
